refactor: 拆分 OpenAI Responses Tool Loop
